refactor(slack,autopilot): split Slack and Autopilot modules for maintainability
- Split slackService into focused modules behind stable facade
- Extract route auth helpers for Autopilot and Slack admin access
- Consolidate Slack Web API calls in dedicated client
- Split Slack socket responsibilities into lifecycle, locks, utils, handlers
- Add comprehensive tests for new modules
- Increase overall test coverage to 95.03%
