+++ apps/web/src/lib/autopilot/route-auth.ts
@@ -1,0 +1,13 @@
+import { userService } from '@/lib/services'
+
+export async function resolveAutopilotWorkspaceUserId(
+  slug: string,
+  contextUser: { id: string; slug: string },
+): Promise<string | null> {
+  if (contextUser.slug === slug) {
+    return contextUser.id
+  }
+
+  const owner = await userService.findIdBySlug(slug)
+  return owner?.id ?? null
+}
