+++ apps/web/src/lib/services/__tests__/autopilot.test.ts
@@ -422,6 +422,35 @@ describe('autopilotService', () => {
       const findCall = mockPrisma.autopilotTask.findFirst.mock.calls[0][0]
       expect(findCall.where).not.toHaveProperty('userId')
     })
+
+    it('can explicitly use the per-user concurrency policy for immediate claims', async () => {
+      const task = { id: 'task-1', leaseOwner: null, leaseExpiresAt: null }
+      mockPrisma.autopilotTask.findFirst.mockResolvedValue(task)
+      mockPrisma.autopilotTask.updateMany.mockResolvedValue({ count: 1 })
+
+      const { claimTaskForImmediateRun } = await import('../autopilot')
+      await claimTaskForImmediateRun({
+        concurrencyPolicy: 'per_user',
+        id: 'task-1',
+        leaseMs: 60_000,
+        leaseOwner: 'worker-1',
+        now,
+      })
+
+      expect(mockPrisma.autopilotTask.findFirst).toHaveBeenCalledWith(
+        expect.objectContaining({
+          where: expect.objectContaining({
+            user: {
+              autopilotTasks: {
+                none: {
+                  leaseExpiresAt: { gt: now },
+                },
+              },
+            },
+          }),
+        }),
+      )
+    })
   })
 
   // -----------------------------------------------------------------------
