+++ apps/web/src/lib/services/slack/errors.ts
@@ -1,0 +1,8 @@
+export function isUniqueConstraintError(error: unknown): error is { code: string } {
+  return (
+    typeof error === 'object' &&
+    error !== null &&
+    'code' in error &&
+    (error as { code?: unknown }).code === 'P2002'
+  )
+}
