+++ apps/web/src/lib/services/slack/text.ts
@@ -1,0 +1,4 @@
+export function normalizeOptionalSlackText(value: string | null | undefined): string | null {
+  const trimmed = value?.trim() ?? ''
+  return trimmed ? trimmed : null
+}
